+++ src/Exception/ErrorException.php
@@ -4,7 +4,7 @@
  * Licensed under The MIT License
  * For full copyright and license information, please see the LICENSE file.
  * Redistributions of files must retain the above copyright notice.
- * 
+ *
  * @copyright (c) 2014, Achmad F. Ibrahim
  * @link https://github.com/Achsoft
  * @license http://opensource.org/licenses/mit-license.php The MIT License (MIT)
@@ -13,14 +13,15 @@
 namespace Achsoft\Component\ErrorHandler\Exception;
 
 /**
- * ErrorException wraper.
+ * Base ErrorException class.
+ *
+ * Thrown on E_ERROR error. This class wraps \ErrorException.
  *
  * @author Achmad F. Ibrahim <acfatah@gmail.com>
- * @package Achsoft\Component\ErrorHandler
- * @version 0.1.1
- * @since 0.1.0
+ * @package Achsoft\Component\ErrorHandler\Exception
+ * @version 0.2.1
  */
 class ErrorException extends \ErrorException
 {
-    
+
 }