+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.1 (2014.07.11)
+* Add *setErrorReporting* method
+* Display error directive set to false
+
+## 0.2.0 (2014.07.08) - BC Breaks
+* Add exception types based on E_* error types
+* Remove constructor
+* Remove default error exception and its setter
+
 ## 0.1.1 (2014.07.05)
 * Add a restore method to restore default error handler